+++ pkg/config/config.go
@@ -41,8 +41,9 @@ var (
 )
 
 const (
-	configMapSchema    = "schema"
-	configMapPlainHttp = "plainHttp"
+	configMapSchema      = "schema"
+	configMapPlainHttp   = "plainHttp"
+	configMapInsecureTls = "insecureTls"
 )
 
 type EndpointSchema string
@@ -60,7 +61,9 @@ type HelmRepositoryData struct {
 	// Schema describes the way how clients communicate with the Helm registry endpoint.
 	Schema EndpointSchema `json:"schema" yaml:"schema"`
 	// PlainHttp indicates that the repository endpoint should be accessed using plain http
-	PlainHttp bool `json:"plainHttp,omitempty" yaml:"plainHttp,omitempty"`
+	PlainHttp   bool `json:"plainHttp,omitempty" yaml:"plainHttp,omitempty"`
+        // InsecureTls allows invalid or selfsigned certificates to be used. This option may be overridden by PlainHttp which forces HTTP traffic.
+	InsecureTLS bool `json:"insecureTls" yaml:"insecureTls"`
 }
 
 // URL returns the full URL Helm repository endpoint including schema.
@@ -154,6 +157,13 @@ func NewHelmRepoDataFromCluster(ctx context.Context, configMapClient configMapIn
 			return nil, fmt.Errorf("failed to parse field %s from configMap %s", configMapPlainHttp, helmRepositoryConfigMapName)
 		}
 	}
+	insecureTls := false
+	if insecureTlsStr, exists := configMap.Data[configMapInsecureTls]; exists {
+		insecureTls, err = strconv.ParseBool(insecureTlsStr)
+		if err != nil {
+			return nil, fmt.Errorf("failed to parse field %s from configMap %s", configMapInsecureTls, helmRepositoryConfigMapName)
+		}
+	}
 	var schema string
 	var schemaExists bool
 	if schema, schemaExists = configMap.Data[configMapSchema]; schemaExists {
@@ -163,9 +173,10 @@ func NewHelmRepoDataFromCluster(ctx context.Context, configMapClient configMapIn
 	}
 
 	repoData := &HelmRepositoryData{
-		Endpoint:  configMap.Data["endpoint"],
-		Schema:    EndpointSchema(schema),
-		PlainHttp: plainHttp,
+		Endpoint:    configMap.Data["endpoint"],
+		Schema:      EndpointSchema(schema),
+		PlainHttp:   plainHttp,
+		InsecureTLS: insecureTls,
 	}
 
 	err = repoData.validate()